logout errorFix, ImageProcessor add, appIcon
